make quake2 architecture independent. from Jaq (Jamie Wilkinson)

+++ client/cl_scrn.c
@@ -614,11 +614,11 @@ int entitycmpfnc( const entity_t *a, const entity_t *b )
 	*/
 	if ( a->model == b->model )
 	{
-		return ( ( int ) a->skin - ( int ) b->skin );
+		return (long int) a->skin - (long int) b->skin;
 	}
 	else
 	{
-		return ( ( int ) a->model - ( int ) b->model );
+		return (long int) a->model - (long int) b->model;
 	}
 }
